+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - feature: install aider
 - feature: install bruno
 - feature: install cline
+- feature: install continue
 
 ## [1.4.0] - 2026-07-24